tests/mediacapture-depth/MediaStream-MediaElement-srcObject-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Assigning mediastream to a video element</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#mediastreams-in-media-elements">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<meta name="flags" content="interact">

<h2>Preconditions</h2>
<ol>
  <li>
    Use a test device with depth camera(embedded or external).
  </li>
  <li>
    When prompted, accept to share your depth stream.
  </li>
</ol>

<div id="log"></div>

<video id="vid"></video>
<script>

async_test(t => {
  let vid = document.getElementById("vid");
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}}})
  .then(stream => {
    let testOncePlaying = () => {
      assert_equals(vid.played.length, 1, "A MediaStream's timeline always consists of a single range");
      assert_equals(vid.played.start(0), 0, "A MediaStream's timeline always consists of a single range");
      assert_equals(vid.played.end(0), vid.currentTime, "A MediaStream's timeline always consists of a single range");
      assert_equals(vid.readyState, vid.HAVE_ENOUGH_DATA, "Upon selecting a media stream, the UA sets readyState to HAVE_ENOUGH_DATA");
      let time = vid.currentTime;
      assert_throws("INVALID_STATE_ERR", () => {
        vid.currentTime = 0;
      });
      assert_equals(vid.currentTime, time, "The UA MUST ignore attempts to set the currentTime attribute");
      vid.removeEventListener("timeupdate", testOncePlaying, false);
      t.done();
    }
    vid.addEventListener("timeupdate", t.step_func(testOncePlaying), false);
    vid.srcObject = stream;
    vid.play();
    assert_true(!vid.seeking, "A MediaStream is not seekable");
    assert_equals(vid.seekable.length, 0, "A MediaStream is not seekable");
    assert_equals(vid.defaultPlaybackRate, 1, "playback rate is always 1");
    assert_equals(vid.playbackRate, 1, "playback rate is always 1");
    assert_equals(vid.buffered.length, 0, "A MediaStream cannot be preloaded.  Therefore, there is no buffered timeranges");
  });
}, "Tests that a MediaStream can be assigned to a video element with srcObject", {timeout: 10000});

</script>
